Statically link to Qt based dependent librarys.
Some of the library dependencies of the ICD Bearer Management plugin use
Qt. We include them in src/3rdparty and link against them statically to
avoid the circular build dependency. Libraries that do not use Qt are
still linked dynamically. The libraries which are statically linked are
libconninet and libgq-gconf.

Task-number: QTBUG-12354