Add the ability to enable various SSL bug workarounds.
There are lots of buggy SSL servers around and to connect to them you
need to disable various features. This commit adds the ability to
disable the SSL ticket extension, the ability to disable the insertion
of empty fragments, and the ability to disable compression.

Task-number: QTBUG-21906

Change-Id: I3e1d0347a46e9030b889bbf15b2aad19b8513b73
Merge-request: 68
Reviewed-by: Peter Hartmann <peter.hartmann@nokia.com>
(cherry picked from commit 78d02e93aca5325fc5be9bfd275862795207abaa)

(commit was cherry-picked from Qt 5 to 4.8 after agreeing with the
author because the merge request was filed against Qt 5.)